refactor: extract transform classes from sync_manifest to transforms.py
Refs: #89
